+++ packages/std/src/math/int64.rs
@@ -13,7 +13,7 @@ use crate::{
     __internal::forward_ref_partial_eq,
 };
 
-use super::conversion::{forward_try_from, try_from_int_to_int};
+use super::conversion::{forward_try_from, try_from_int_to_int, wrapped_int_to_primitive};
 use super::impl_int_serde;
 use super::num_consts::NumConsts;
 
@@ -339,6 +339,10 @@ impl From<i8> for Int64 {
     }
 }
 
+// Int to int
+wrapped_int_to_primitive!(Int64, i64);
+wrapped_int_to_primitive!(Int64, i128);
+
 // Int to Int
 try_from_int_to_int!(Int128, Int64);
 try_from_int_to_int!(Int256, Int64);
@@ -597,6 +601,15 @@ mod tests {
         let num2 = Int64::from_le_bytes(le_bytes);
         assert_eq!(num1, Int64::from(65536u32 + 512 + 3));
         assert_eq!(num1, num2);
+    }
+
+    #[test]
+    fn int64_convert_to() {
+        let a = Int64::new(5);
+        assert_eq!(i64::from(a), 5);
+
+        let a = Int64::new(5);
+        assert_eq!(i128::from(a), 5);
     }
 
     #[test]