+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+- cosmwasm-std: Implement `From<Uint64> for u{64,128}`,
+  `From<Uint128> for u128`, `From<Int64> for i{64,128}`, and
+  `From<Int128> for i128` ([#2268])
+
+[#2268]: https://github.com/CosmWasm/cosmwasm/issues/2268
+
 ## [2.2.0] - 2024-12-17
 
 ### Added